+++ lib/usage.sh
@@ -94,7 +94,7 @@ Informational Arguments:
 
 Remote Transfer Arguments:
       --sftp SERVER
-                    Transfer the output file to remote SFTP server.
+                    Transfer the output and log file to remote SFTP server.
                     SERVER must be specified in the form [user@]host:[path]
       --sftp-port PORT
                     Remote SFTP server port (default: 22).
@@ -105,7 +105,7 @@ Remote Transfer Arguments:
                     Allow setting SSH options as key=value pairs.
                     Can be used multiple times to set multiple options.
       --s3-provider
-                    Transfer the output and log files to S3 service.
+                    Transfer the output and log file to S3 service.
                     Options: amazon, google, ibm
       --s3-region
                     S3 region name (default: us-east-1 [amazon], us-south [ibm]).
@@ -124,10 +124,7 @@ Remote Transfer Arguments:
                     Transfer the log file to AWS S3 using a pre-signed URL.
                     Use single quotes to enclose the URL.
       --azure-storage-sas-url URL
-                    Transfer the output file to Azure Storage using a SAS URL.
-                    Use single quotes to enclose the URL.
-      --azure-storage-sas-url-log-file URL
-                    Transfer the log file to Azure Storage using a SAS URL.
+                    Transfer the output and log file to Azure Storage using a SAS URL.
                     Use single quotes to enclose the URL.
       --delete-local-on-successful-transfer
                     Delete local output and log files on successful transfer.